Refactor schedule-related constants and types for better organization
