+++ src/lib/scheduleMatrixConstants.tsx
@@ -1,0 +1,86 @@
+// Mapeo de franjas horarias
+export const TIME_SLOTS = [
+  "08:20",
+  "09:40",
+  "11:00",
+  "12:20",
+  "14:50",
+  "16:10",
+  "17:30",
+  "18:50",
+  "20:10",
+];
+
+// Rango de duración de los módulos
+export const TIME_RANGES = {
+  "08:20": "09:30",
+  "09:40": "10:50",
+  "11:00": "12:10",
+  "12:20": "13:30",
+  "14:50": "16:00",
+  "16:10": "17:20",
+  "17:30": "18:40",
+  "18:50": "20:00",
+  "20:10": "21:20",
+};
+
+// Días de la semana (incluye sábado)
+export const DAYS = ["L", "M", "W", "J", "V", "S"];
+
+// Mapeo de bloques (L=Lunes, M=Martes, W=Miércoles, J=Jueves, V=Viernes, S=Sábado)
+export const BLOCK_MAP: Record<string, { dia: string; hora: string }> = {
+  l1: { dia: "L", hora: "08:20" },
+  l2: { dia: "L", hora: "09:40" },
+  l3: { dia: "L", hora: "11:00" },
+  l4: { dia: "L", hora: "12:20" },
+  l5: { dia: "L", hora: "14:50" },
+  l6: { dia: "L", hora: "16:10" },
+  l7: { dia: "L", hora: "17:30" },
+  l8: { dia: "L", hora: "18:50" },
+  l9: { dia: "L", hora: "20:10" },
+  m1: { dia: "M", hora: "08:20" },
+  m2: { dia: "M", hora: "09:40" },
+  m3: { dia: "M", hora: "11:00" },
+  m4: { dia: "M", hora: "12:20" },
+  m5: { dia: "M", hora: "14:50" },
+  m6: { dia: "M", hora: "16:10" },
+  m7: { dia: "M", hora: "17:30" },
+  m8: { dia: "M", hora: "18:50" },
+  m9: { dia: "M", hora: "20:10" },
+  w1: { dia: "W", hora: "08:20" },
+  w2: { dia: "W", hora: "09:40" },
+  w3: { dia: "W", hora: "11:00" },
+  w4: { dia: "W", hora: "12:20" },
+  w5: { dia: "W", hora: "14:50" },
+  w6: { dia: "W", hora: "16:10" },
+  w7: { dia: "W", hora: "17:30" },
+  w8: { dia: "W", hora: "18:50" },
+  w9: { dia: "W", hora: "20:10" },
+  j1: { dia: "J", hora: "08:20" },
+  j2: { dia: "J", hora: "09:40" },
+  j3: { dia: "J", hora: "11:00" },
+  j4: { dia: "J", hora: "12:20" },
+  j5: { dia: "J", hora: "14:50" },
+  j6: { dia: "J", hora: "16:10" },
+  j7: { dia: "J", hora: "17:30" },
+  j8: { dia: "J", hora: "18:50" },
+  j9: { dia: "J", hora: "20:10" },
+  v1: { dia: "V", hora: "08:20" },
+  v2: { dia: "V", hora: "09:40" },
+  v3: { dia: "V", hora: "11:00" },
+  v4: { dia: "V", hora: "12:20" },
+  v5: { dia: "V", hora: "14:50" },
+  v6: { dia: "V", hora: "16:10" },
+  v7: { dia: "V", hora: "17:30" },
+  v8: { dia: "V", hora: "18:50" },
+  v9: { dia: "V", hora: "20:10" },
+  s1: { dia: "S", hora: "08:20" },
+  s2: { dia: "S", hora: "09:40" },
+  s3: { dia: "S", hora: "11:00" },
+  s4: { dia: "S", hora: "12:20" },
+  s5: { dia: "S", hora: "14:50" },
+  s6: { dia: "S", hora: "16:10" },
+  s7: { dia: "S", hora: "17:30" },
+  s8: { dia: "S", hora: "18:50" },
+  s9: { dia: "S", hora: "20:10" },
+};
